refactor: delete useless sections, improved readability
