+++ task18tests/CommandSchedulerTests.cs
@@ -5,18 +5,8 @@ using ICommand;
 using RoundRobinScheduler;
 using System.Threading;
 using System.Diagnostics;
-using ISteppable;
 
 namespace task18tests;
-
-public class StepCommand : ICommand.ICommand, ISteppable.ISteppable
-{
-    public int StepsLeft;
-    public int Executed = 0;
-    public StepCommand(int steps) { StepsLeft = steps; }
-    public void Execute() { if (StepsLeft > 0) { StepsLeft--; Executed++; } }
-    public bool IsDone => StepsLeft == 0;
-}
 
 public class CommandSchedulerTests
 {
@@ -31,14 +21,12 @@ public class CommandSchedulerTests
         cs.EnqueueCommand(cmd2);
         while (!cmd1.IsDone || !cmd2.IsDone) Thread.Sleep(10);
         cs.Stop();
-        Assert.Equal(5, cmd1.Executed);
-        Assert.Equal(5, cmd2.Executed);
-        Assert.InRange(cmd1.Executed, 5, 5);
-        Assert.InRange(cmd2.Executed, 5, 5);
+        Assert.True(cmd1.Executed == 5);
+        Assert.True(cmd2.Executed == 5);
     }
 
     [Fact]
-    public void ShortCommand_NotBlockedByLong()
+    public void OneShortOneLongCommands_ShortNotBlockedByLong()
     {
         var scheduler = new RoundRobinScheduler.RoundRobinScheduler();
         var cs = new CommandScheduler(scheduler);
@@ -49,10 +37,13 @@ public class CommandSchedulerTests
         while (!shortCmd.IsDone) Thread.Sleep(5);
         cs.Stop();
         Assert.True(shortCmd.IsDone);
+        Assert.True(shortCmd.Executed == 1);
+        Assert.True(longCmd.Executed == 100);
+        Assert.True(longCmd.IsDone);
     }
 
     [Fact]
-    public void NoWork_DoesNotWasteCpu()
+    public void NoCommands_CommandSchedulerCorrectlyWaitsForCommands()
     {
         var scheduler = new RoundRobinScheduler.RoundRobinScheduler();
         var cs = new CommandScheduler(scheduler);
@@ -60,16 +51,6 @@ public class CommandSchedulerTests
         Thread.Sleep(100);
         cs.Stop();
         Assert.True(sw.ElapsedMilliseconds >= 100);
-    }
-
-    [Fact]
-    public void NoCommands_ThreadStopsGracefully()
-    {
-        var scheduler = new RoundRobinScheduler.RoundRobinScheduler();
-        var cs = new CommandScheduler(scheduler);
-        Thread.Sleep(50);
-        cs.Stop();
-        Assert.True(true);
-    }
+    } 
 }
